feat: wallet tab
