+++ src/frontend/src/lib/i18n/en.json
@@ -90,7 +90,8 @@
 		"credits": "Credits",
 		"transactions": "Transactions",
 		"transfer_icp_info": "Find more information about transferring funds to your mission control in the",
-		"transfer_to_account_identifier": "To fund your mission control, you can send ICP to its account identifier"
+		"transfer_to_account_identifier": "To fund your mission control, you can send ICP to its account identifier",
+		"wallet": "Wallet"
 	},
 	"authentication": {
 		"title": "Authentication",
